measurement: fix SACCH interval remapping in test_is_meas_overdue()
The function test_is_meas_overdue() checks if a SACCH block was lost in
order to be able to conclude the measurement interval even when the
SAACH block gets lost.

Unfortunately, the measurements for the SACCH do not arrive in order with
the normal voice measurements. The measurement for the SACCH is shifted
by -26 frames but arrives with the correct frame number. The detection
logic that checks for dropped out SACCH frames does not take this
correctly into account. It calculates the frame number shift, but it
still expects the measurements to arrive in order. This needs to be
fixed.

- Add logic to work around the frame number shift
- Adapt unit tests

Change-Id: Idf6db73f67b7273c3d5965616c9e78f66b74a406
Related: OS#3502